src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/reference/SingleShortAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.attribute.SingleAttributeParser;
import org.openqa.selenium.WebElement;

public class SingleShortAttributeParser extends SingleAttributeParser {

    public SingleShortAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Short.parseShort(element.getAttribute(super.attribute));
    }

}
