src/test/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/reference/SingleShortAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class SingleShortAttributeParserTest{

    @Test
    public void constructorTest() {
        String attribute = "";
        SingleShortAttributeParser parser = new SingleShortAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        SingleShortAttributeParser parser = new SingleShortAttributeParser("");

        assertEquals((short)2, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "2";
            }
        }));
    }

}
